Flipping the burgers
This was exactly backwards. The "channels" label contains the list of possible channels for your product - which would be 'stable,alpha'. 
The "default" label should contain a single channel name, in this case 'alpha'.

Please try flipping those definitions around and you should get a better result.

Thanks and regards,