feat: Onboarding MariaDB to FBC
The `Makefile` is the official one recommended by the [docs](https://redhat-openshift-ecosystem.github.io/operator-pipelines/users/fbc_onboarding/).

Ref: https://olm.operatorframework.io/docs/reference/catalog-templates/#semver-template

Semver was not used intentionally as it will differ from our desired LTS strategy. However it is important to note that:

The [recommended](https://redhat-openshift-ecosystem.github.io/operator-pipelines/users/fbc_workflow/) template from maintainability point
of view is SemVer. And **IF** the documentation is to be trusted, in the future if we want a single step PR process instead of 2, we'll have
to use `SemVer`.

Catalogs this time were generated with `make catalogs`, however in the future, they will be generated by a `release-config.yaml`.

`make fbc-onboarding` was ran once to generate the needed structure. It generated a bunch of "v4.x.yaml" files, which were all removed in
favor of one that will be used for all versions that we support (16-19). In the future if something irregular happens and we need more of
them, they can easily be added.

This file in the future will also be handled by the `release-config.yaml`.

Read: https://redhat-openshift-ecosystem.github.io/operator-pipelines/users/fbc_autorelease/

Release configs automate the process of generating the `catalogs`. In the future, this file must be generated in new releases to point where
they should go.

It is not clear if we still need to specify these in the annotations and the ClusterServiceVersion... My assumption is that the others are
ignored and only what is is generated in the catalogs matter.